fixed a bug in the tracking of the page views. It's also simpler now.
Basically we insert the number of page views directly in the page
instead of storing it in a separate cookie. What was happening before
is that the browsers kept reading the old page views cookies, so they
were stuck showing whatever number of page views they've seen when the
first visited the page.